fix: don't remove workspaces when a kcp job dont succeed (#222)
* remove kcp workspaces only if job succeed

* remove kcp workspaces only if job succeed

+++ scripts/install-appstudio-kcp.sh
@@ -46,8 +46,9 @@ export CI=${CI:-"false"}
 export TIMEOUT_PID=
 
 function catchFinish() {
-    # Remove workspaces only if we are in CI
-    if [[ $CI != "false" ]]
+    local RESULT=$?
+    # Remove workspaces only if we are in CI and the job was succeded
+    if [[ "$CI" != "false" ]] && [[ "$RESULT" == "0" ]]
     then
         kubectl kcp workspace use '~' || true
         kubectl delete ws "${APPSTUDIO_WORKSPACE}" "${HACBS_WORKSPACE}" "${USER_APPSTUDIO_WORKSPACE}" "${COMPUTE_WORKSPACE}" "${USER_HACBS_WORKSPACE}" || true
